# feat(perf-detector-threshold-configuration) Detector dry runs and new thresholds for extended detectors.