--- a/src/sentry/utils/performance_issues/performance_detection.py
+++ b/src/sentry/utils/performance_issues/performance_detection.py
@@ -272,7 +272,7 @@ def get_detection_settings(project_id: Optional[int] = None) -> Dict[DetectorTyp
             "detection_enabled": settings["n_plus_one_api_calls_detection_enabled"],
         },
         DetectorType.N_PLUS_ONE_API_CALLS_EXTENDED: {
-            "total_duration": 500,  # ms
+            "total_duration": 300,  # ms
             "concurrency_threshold": 5,  # ms
             "count": 10,
             "allowed_span_ops": ["http.client"],
@@ -300,8 +300,8 @@ def get_detection_settings(project_id: Optional[int] = None) -> Dict[DetectorTyp
             "detection_enabled": settings["consecutive_http_spans_detection_enabled"],
         },
         DetectorType.CONSECUTIVE_HTTP_OP_EXTENDED: {
-            # time saved by running all queries in parallel
-            "min_time_saved": 2000,
+            "span_duration_threshold": 500,  # ms
+            "min_time_saved": 2000,  # time saved by running all queries in parallel
             "consecutive_count_threshold": 3,
             "max_duration_between_spans": 1000,  # ms
         },
@@ -312,6 +312,29 @@ def get_detection_settings(project_id: Optional[int] = None) -> Dict[DetectorTyp
         DetectorType.HTTP_OVERHEAD: {
             "http_request_delay_threshold": settings["http_request_delay_threshold"],
             "detection_enabled": settings["http_overhead_detection_enabled"],
+        },
+    }
+
+
+#
+def get_dry_run_detection_settings(project_id: Optional[int] = None) -> Dict[DetectorType, Any]:
+    settings = get_merged_settings(project_id)
+
+    return {
+        DetectorType.N_PLUS_ONE_DB_QUERIES: {
+            "count": settings["n_plus_one_db_count"],
+            "duration_threshold": 50,  # ms
+            "detection_enabled": settings["n_plus_one_db_queries_detection_enabled"],
+        },
+        DetectorType.UNCOMPRESSED_ASSETS: {
+            "size_threshold_bytes": settings["uncompressed_asset_size_threshold"],
+            "duration_threshold": 300,  # ms
+            "allowed_span_ops": ["resource.css", "resource.script"],
+            "detection_enabled": settings["uncompressed_assets_detection_enabled"],
+        },
+        DetectorType.LARGE_HTTP_PAYLOAD: {
+            "payload_size_threshold": 500000,  # kb
+            "detection_enabled": settings["large_http_payload_detection_enabled"],
         },
     }
 
@@ -346,6 +369,22 @@ def _detect_performance_problems(
 
     # Metrics reporting only for detection, not created issues.
     report_metrics_for_detectors(data, event_id, detectors, sdk_span, project.organization)
+
+    # TODO Abdullah Khan: Remove code after dry run evaluating changes in detection ---------------------
+    detection_dry_run_settings = get_dry_run_detection_settings(project_id)
+    dry_run_detectors: List[PerformanceDetector] = [
+        NPlusOneDBSpanDetector(detection_dry_run_settings, data),
+        UncompressedAssetSpanDetector(detection_dry_run_settings, data),
+        LargeHTTPPayloadDetector(detection_dry_run_settings, data),
+    ]
+
+    for dry_run_detector in dry_run_detectors:
+        run_detector_on_data(dry_run_detector, data)
+
+    report_metrics_for_detectors(
+        data, event_id, dry_run_detectors, sdk_span, project.organization, True
+    )
+    # ----------------------------------------------------------------------------------------------------
 
     organization = cast(Organization, project.organization)
     if project is None or organization is None:
@@ -401,6 +440,7 @@ def report_metrics_for_detectors(
     detectors: Sequence[PerformanceDetector],
     sdk_span: Any,
     organization: Organization,
+    is_dry_run: bool = False,
 ):
     all_detected_problems = [i for d in detectors for i in d.stored_problems]
     has_detected_problems = bool(all_detected_problems)
@@ -450,6 +490,10 @@ def report_metrics_for_detectors(
         "sdk_name": sdk_name,
         "is_early_adopter": organization.flags.early_adopter.is_set,
     }
+
+    if is_dry_run:
+        detected_tags["is_dry_run"] = True
+
     event_integrations = event.get("sdk", {}).get("integrations", []) or []
 
     for integration_name in INTEGRATIONS_OF_INTEREST: