1.  Added min span duration threshold to extended Consecutive HTTP span detector. With default of 500ms.
2. Reduced total duration threshold of extended N+1 API Calls issue to 300ms.
3. Added dry runs for the following:

- **N+1 db**: Dry run detection with 50ms total duration threshold.
- **Large HTTP Payload**: Dry run detection with 300kb size threshold.
- **Uncompressed asset**: Dry run detection with default duration of  300ms.